Integrate new TsWriter interfaces into the TV library. Much of the integration itself is complete. However there are still lots of compilation errors outside the library (due to changes which I haven't yet completed/committed) and the EPG grabbing classes in particular are not integrated properly with the service.

+++ TvEngine3/Mediaportal/TV/Server/TVLibrary/Interfaces/Analyzer/IGrabberSiMpeg.cs
@@ -91,7 +91,7 @@ namespace Mediaportal.TV.Server.TVLibrary.Interfaces.Analyzer
     /// <returns><c>true</c> if the program's details are successfully retrieved, otherwise <c>false</c></returns>
     [PreserveSig]
     [return: MarshalAs(UnmanagedType.I1)]
-    bool GetProgramByIndex(uint index,
+    bool GetProgramByIndex(ushort index,
                             out ushort programNumber,
                             out ushort pmtPid,
                             [MarshalAs(UnmanagedType.I1)] out bool isPmtReceived,
